Fixes #3. Reorganised sty file.
Fixes #3: the issue was that /orm/uniqueness bar set roles sets the
/pgf/minimum width, which is used later in \pgfmovepathto and friends.
The minimum width is not affected by scale, while the path commands are.
This fix extracts the scale from \pgfgettransformentries and sets the
/pgf/minimum width accordingly through /orm/uniqueness bar set roles.

Reorganised sty file: power & sequence types were in the middle of the
uniqueness code; moved to the bottom.